requirements.txt
aiohttp==3.8.3
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
Flask==2.2.2
frozenlist==1.3.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pyaes==1.6.1
pyasn1==0.4.8
pyparsing==3.0.9
pyTelegramBotAPI==4.7.1
python-dotenv==0.21.0
redis==4.3.4
requests==2.28.1
rsa==4.9
Telethon==1.25.4
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.10.0
